backend/requirements.txt
google-generativeai
fastapi
uvicorn[standard]
python-dotenv
PyYAML
reportlab
